docs(readme): strengthen portfolio presentation and document warehouse architecture
