+++ .gitignore
@@ -7,6 +7,7 @@ desktop.ini
 .idea/
 .vscode/
 *.code-workspace
+.claude/
 
 # Python bytecode and caches
 __pycache__/
@@ -53,7 +54,8 @@ temp/
 
 # Local analytical storage
 local_lake/
-warehouse/
+warehouse/*
+!warehouse/.gitkeep
 *.duckdb
 *.duckdb.wal
 *.db
