+++ data/sample/cloud_cost_usage_sample_manifest.json
@@ -1,0 +1,10 @@
+{
+  "batch_id": "batch-20260406T184645Z",
+  "run_date": "2026-04-06",
+  "row_count": 613,
+  "start_date": "2026-03-08",
+  "end_date": "2026-04-06",
+  "file_format": "parquet",
+  "data_file": "C:\\Users\\bruno\\Desktop\\Projetos\\finops-cost-capitalization-pipeline\\local_lake\\raw\\cloud_costs\\run_date=2026-04-06\\cloud_cost_usage.parquet",
+  "sample_file": "C:\\Users\\bruno\\Desktop\\Projetos\\finops-cost-capitalization-pipeline\\data\\sample\\cloud_cost_usage_sample.csv"
+}
