+++ dbt/models/bronze/brz_cloud_cost_usage.sql
@@ -1,0 +1,37 @@
+{{ config(materialized="table") }}
+
+select
+    line_item_id,
+    billing_period_start,
+    usage_start_time,
+    usage_end_time,
+    invoice_date,
+    cloud_provider,
+    payer_account_id,
+    linked_account_id,
+    service,
+    usage_type,
+    operation,
+    region,
+    resource_id,
+    line_item_type,
+    usage_amount,
+    currency_code,
+    unblended_cost,
+    blended_cost,
+    discount_type,
+    cost_center,
+    owner_team,
+    environment,
+    product_line,
+    capitalization_candidate,
+    initiative_id,
+    initiative_stage,
+    asset_lifecycle,
+    accounting_policy_version,
+    generator_batch_id,
+    generated_at,
+    filename as source_file_name,
+    regexp_extract(filename, 'run_date=([0-9]{4}-[0-9]{2}-[0-9]{2})', 1) as run_date,
+    current_timestamp as ingestion_timestamp
+from read_parquet('{{ var("raw_cloud_cost_glob") }}', union_by_name = true, filename = true)
