+++ dbt/models/bronze/schema.yml
@@ -1,0 +1,37 @@
+version: 2
+
+models:
+  - name: brz_cloud_cost_usage
+    description: >
+      Bronze ingestion table that reads raw billing files from the local mock
+      lake and preserves the provider-like structure with ingestion metadata.
+    columns:
+      - name: line_item_id
+        description: Synthetic raw billing line identifier.
+        tests:
+          - not_null
+      - name: source_file_name
+        description: File path reported by DuckDB during Parquet ingestion.
+        tests:
+          - not_null
+      - name: generator_batch_id
+        description: Batch identifier created by the synthetic generator.
+        tests:
+          - not_null
+      - name: run_date
+        description: Raw partition date inferred from the file path.
+        tests:
+          - not_null
+
+  - name: brz_generation_manifest
+    description: Manifest metadata for each generated raw landing partition.
+    columns:
+      - name: batch_id
+        description: Synthetic batch identifier.
+        tests:
+          - not_null
+          - unique
+      - name: row_count
+        description: Number of rows described by the manifest.
+        tests:
+          - not_null
