+++ requirements-dev.txt
@@ -1,4 +1,1 @@
--r requirements.txt
-pytest>=8.0,<9.0
-pytest-cov>=5.0,<6.0
-ruff>=0.5,<0.8
+-e .[dev]
