--- requirements.txt
@@ -1,5 +1,0 @@
-pandas>=2.2,<3.0
-numpy>=2.0,<3.0
-faker>=25.0,<30.0
-pyarrow>=16.0,<20.0
-pyyaml>=6.0,<7.0
