+++ src/finops_capex/cli/__init__.py
@@ -1,0 +1,1 @@
+"""CLI entrypoints for repository workflows."""
