+++ tests/conftest.py
@@ -5,7 +5,6 @@ from __future__ import annotations
 import sys
 from pathlib import Path
 
-
 REPOSITORY_ROOT = Path(__file__).resolve().parents[1]
 SRC_PATH = REPOSITORY_ROOT / "src"
